fix: some test mocks moved
